server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.142.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.1

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
